requirements.txt
# requirements.txt

django
gunicorn
mysqlclient
django-environ
boto3
django-storages